refactor: 최근 조회 목록 조회에서 projection으로 조회하던 것을 entity로 조회하게끔 변경

+++ src/main/java/kr/kro/airbob/domain/accommodation/dto/AccommodationResponse.java
@@ -58,7 +58,7 @@ public class AccommodationResponse {
 		String accommodationName,
 		String thumbnailUrl,
 		List<AmenityInfoResponse> amenities,
-		Double averageRating,
+		BigDecimal averageRating,
 		Boolean isInWishlist
 	) {
 	}
